Move functions out of math module and delete it
